improvement on dhcp_server

+++ dhcp/dhcp_server/delete.py
@@ -8,24 +8,28 @@ import subprocess
 
 interface = sys.argv[1]
 
-# -------------------------- check or write  -----------------------------
+# --------------------------  check and replace  -----------------------------
 # function to check if the file contain the value, if it isn't there the function will write it
-def cow(file_dir,data):
-     with open(file_dir,'r+') as input_file:
-         lines = [line.strip().replace('\n','') for line in input_file.readlines()]
-         if data in lines:
-            print("data found")
-            input_file.write('#DELETED')
+def cor(file_dir,data,replace):
+    with fileinput.FileInput(file_dir,inplace=True) as file:
+        for line in file:
+            print(line.replace(data,replace),end=' ')
 # ------------------------------------------------------------------------
 def main(interface):
-
-    cow( '/etc/default/isc-dhcp-server', 'INTERFACESv4=\"' + interface + '\"' )
-    cow( '/etc/dhcpcd.conf' , 'include \"/etc/dhcpcd.d/' + interface + '.conf\"' + ';' )
-    cow( '/etc/network/interfaces' , 'source /etc/network/interfaces.d/' + interface )
+    cor( '/etc/default/isc-dhcp-server', 'INTERFACESv4=\"' + interface + '\"', ' ' )
+    cor( '/etc/dhcpcd.conf' , 'include \"/etc/dhcpcd.d/' + interface + '.conf\"' + ';' , ' ' )
+    cor( '/etc/network/interfaces' , 'source /etc/network/interfaces.d/' + interface, ' ' )
 
     os.remove('/etc/network/interfaces.d/'+interface)
-    os.remove('/etc/dhcpcd.d/'+interface+'.conf')
-    os.remove('/etc/dhcpcd.d/'+interface+'.conf.disabled')
+    dhcp_dir = os.listdir('/etc/dhcpcd.d/')
+    for files in dhcp_dir:
+        if files == interface + '.conf': # check if the file exist in the directory and erase it
+            print('/etc/dhcpcp.d/'+ interface+'.conf exists')
+            os.remove('/etc/dhcpcd.d/'+ interface + '.conf')
+        if files == interface + '.conf.disabled': # check if the file exist in the directory and erase it
+            print('/etc/dhcpcp.d/'+ interface +'.conf.disabled exists')
+            os.remove('/etc/dhcpcd.d/'+ interface + '.conf.disabled')
+
 # ------------------------------------------------------------------------
 
 main(interface)
